requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
drf-spectacular==0.28.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
magic-filter==1.0.12
multidict==6.2.0
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
